pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kimsfinance"
version = "0.1.0"
description = "Blazing-fast financial charting for quantitative trading - 178x faster with GPU acceleration, WebP optimization, and production-ready quality. Built for hedge funds, algo traders, and quants."
readme = "README.md"
requires-python = ">=3.13"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "Kim Asplund", email = "hello@asplund.kim"}
]
maintainers = [
    {name = "Kim Asplund", email = "hello@asplund.kim"}
]
keywords = ["finance", "trading", "gpu", "candlestick", "charts", "visualization", "cuda", "webp", "performance", "quantitative", "quant", "ohlc", "ohlcv", "algorithmic-trading", "technical-analysis", "polars", "numpy", "hedge-fund", "high-frequency"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Natural Language :: English",
    "Typing :: Typed",
    "Operating System :: OS Independent",
]

dependencies = [
    "polars>=1.0",
    "numpy>=2.0",
    "pandas>=2.0",
    "Pillow>=12.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
gpu = [
    "cudf-cu13>=25.12",  # CUDA 13 support (RTX 3500 Ada)
    "cupy-cuda13x>=13.0",  # CUDA 13 support
]
jit = [
    "numba>=0.59",  # Python 3.13 support
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "tox>=4.0",
]
test = [
    "kimsfinance[dev]",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=5.1.0",
    "mplfinance>=0.12.10b0",
    "pyarrow>=15.0",
    "svgwrite>=1.4",
]
all = [
    "kimsfinance[gpu,jit,dev,test]",
]

[project.urls]
Homepage = "https://asplund.kim"
Documentation = "https://github.com/kimasplund/kimsfinance#readme"
Repository = "https://github.com/kimasplund/kimsfinance"
Issues = "https://github.com/kimasplund/kimsfinance/issues"
"Commercial License" = "https://github.com/kimasplund/kimsfinance/blob/master/COMMERCIAL-LICENSE.md"
Funding = "https://github.com/kimasplund/kimsfinance#funding"

[tool.setuptools.packages.find]
where = ["."]
include = ["kimsfinance*"]

[tool.setuptools.package-data]
kimsfinance = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py313', 'py314']

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
